pyproject.toml
[tool.poetry]
name = "carsalon"
version = "0.1.0"
description = ""
authors = ["It Soda <nikshchegilskiy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
djoser = "^2.2.2"
pillow = "^10.1.0"
django-debug-toolbar = "^4.2.0"
django-environ = "0.10.0"
django-extensions = "3.2.3"
django-redis = "5.3.0"
redis = "4.6.0"
celery = {extras = ["redis"], version = "^5.3.6"}
yookassa = "2.4.0"
drf-extensions = "0.7.1"
mysqlclient = "2.1.1"
gunicorn = "^21.2.0"
django-cors-headers = "^4.3.1"
channels = {extras = ["daphne"], version = "^4.0.0"}
channels-redis = "^4.1.0"
daphne = "^4.0.0"
djangochannelsrestframework = "^1.2.0"
python-decouple = "^3.8"
phonenumbers = "^8.13.28"
django-phonenumber-field = "^7.3.0"
twilio = "^8.11.1"
telebot = "^0.0.5"
boto3 = "^1.34.38"

[tool.poetry.group.dev.dependencies]
flake8 = "5.0.4"
isort = "5.10.1"
black = "23.10.1"
coverage = "7.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"